+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncSequence.java
@@ -1,0 +1,12 @@
+package com.github.jlangch.venice.impl.types.collections;
+
+import java.util.List;
+
+import com.github.jlangch.venice.impl.types.VncVal;
+
+
+public abstract class VncSequence extends VncCollection {
+
+	public abstract List<VncVal> getList();
+
+}